ov680: Remove usage of power down signal
BZ: 213750

A firmware change was integrated to correct the problem that the
ISP1_PWDN gpio line was mapped in the gpio table to the wrong gpio.
Now the ov680 platform code actually does set the ISP1_PWDN to high
when the ov680 is not powered on.

However, this caused a large increase in the idle consumption of the
device. Until we understand how the ov680 is supposed to be powered
down correctly, remove the flip of the ISP1_PWDN to high when the
ov680 is not turned on.

Change-Id: Ia5040ce8452213170ab67bcc59259752057e06b5
Signed-off-by: Soren Friis <soren.friis@intel.com>