cdc-acm: Avoid memory corruption in acm driver.
BZ: 215926

There have one race condition if acm get disconnect and application is
trying to do write operation for acm tty device. The acm_tty_write flow
will modify the urb which was already freed in disconnect flow.

Even there have no kernel panic due to acm driver is using kmalloc area
memory. But it maybe impact to other components which is just allocate
the memory from current freed urb.

So we need to check if acm was already get disconnect before submit the
urb in acm_tty_write. But it is still dangerous.

for disconnect flow:
1, set acm->disconnected to true.
2, kill urbs.
3, free urbs.

So before step 1, the acm->disconnected is still false. Then our
condition check will be not valid, and driver will continue to submit
urb. But if urb submit can be finish before step 3, then that is fine.

For this dangerous, this patch will call usb_get_urb to prevent urb get
free in step 3.

Change-Id: I3859621c329f3cb891c4dbd9cc1cf91213845d85
Signed-off-by: Wang, Yu <yu.y.wang@intel.com>