EM/THERMAL/RAPL: Update rapl calculations based on latest SDM
[Description]
When PL1 is set using the sysfs interface exposed by
RAPL/PowerCap framework, the limit set is not accurate
to the closest multiple of the power unit.
The offset is due to the powercap implementation
which is based on the Power unit register
MSR 0x606 (bits [3:0]) description provided in the
old version IA SW developers manual.

[Implementation]
This patch re-structures the way units are computed
and the caluclation for getting real and raw values for
power,energy & time.
This is done referring the latest SDM vol.3B:table:35-7
based on which the actual implementation of the Punit code
for calculating the RAPL related parameters are done.

Change-Id: I0fc483a83b8006261f5e5545cc1a7c489ec6fbc2
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-2297
Signed-off-by: Ajay Thomas <ajay.thomas.david.rajamanickam@intel.com>
Reviewed-on: https://android.intel.com:443/265192