drm/i915: Before presenting a buffer wait for its last write
On the Intel Android HWC we do a wait before setting a plane, to be
sure that the buffer we are about to present is ready. But this wait
was performed by waiting for the last read operation to finish, when
we should instead wait for the last write. The current change exposes
the fuctionality of waiting for the last write to i915 clients that
want to do so.

A second wait for the last read, that should also be a wait for the
last write was present in the SETDISPLAY implementation and needed a
similar fix included with this patch.

This is a fix for a performance drop recently reported on a benchmark.
Note that, when testing, you should apply this fix together with the
needed i915 client (i.e. the Intel Android HWC) change for waiting on
writes.

The forementioned benchmark was uncommon because of two reasons:

1) It was one of those applications that always renders 90 degrees
   rotated (regardless of the current display orientation).

2) It used the EGL_BUFFER_PRESERVED bit when creating the surface,
   meaning that frames other than the first do not start from scratch
   but they can rely on the contents from the previous frame.

When a surface is double buffered the EGL_BUFFER_PRESERVED behaviour
is indeed an illusion that the driver implements for the application.
Our current implementation copies from new front buffer to the new
back buffer at the end of every eglSwapBuffers call.

Apart from knowing a bit about the benchmark, we should be familiar
with the sprite planes from the i915 driver to see how the need for
this patch arised. Let us assume you are familiar with sprite planes.
As you know sprite planes are a good way of compositing the contents
of an OpenGL surface, saving power compared to the fallback
composition path provided by SurfaceFlinger.

However, a composition layer can sometimes have a transform applied
to it, for example when an Android display is rotated 90 degrees, the
composition can have layers needing a 90 degrees transform before
compositing. The main issue with this kind of transform is that it
discards the sprite planes as a valid composition method.

In order to make the sprite planes usable even when there is a
transform, a new feature has been introduced in the OpenGL ES
driver: the pre-rotation.

A pre-rotated surface in one whose contents are generated with a
target rotation (for example 90 degrees). The actual transform to
apply is hinted by the Android EGL at every eglSwapBuffers call, in
such a way that OpenGL always generates the content in the same
orientation the current display has. By having the surface contents
generated with the target orientation, there is no need to apply any
further transform and the sprite planes are able to do the
composition.

But the introduction of the pre-rotation feature soon triggered a new
burning issue: the forementioned benchmark had got a 40% performance
drop, and also a flicker on my VLV2 testing machine.

Was there anything wrong in the OpenGL implementation? Actually not,
what happened was that the OpenGL pre-rotation feature created a new
escenario for the forementioned benchmak. For the first time ever the
EGL surface rendered by the forementiond benchmark could be composed
by means of a sprite plane, unmasking previously unknown issues.

Because the forementioned benchmark relies on the
EGL_BUFFER_PRESERVED bit, on an eglSwapBuffers, two relevant things
happened:

- The new front buffer was queued for SurfaceFlinger to compose a
  window with its contents.

- The GPU commands for a blit from the new front buffer to the back
  buffer were added to the command buffer.

The GPU commands took a while to be executed. In the meantime
SurfaceFlinger asked the HWC to compose the front buffer.

The HWC would then check that the buffer was ready by invoking the
GEM_WAIT ioctl. However, the readiness were waiting for was the
completion of the last read operation involving the buffer. And
because the last read operation was a blit command that had just being
issued, it would stall until the blit finished.

It would also try to stall a second time on the SETDISPLAY ioctl,
waiting for the last read to happen (when again it should have waited
for the last write operation).

The first stall was fixed by adding a flag to the GEM_WAIT call, so
the HWC can explicitly ask for waiting for the last write. The second
stall in SETDISPLAY was fixed in a similar way.

The overall effect of both fixes included in this patch is that the
two wrong stalls are gone. Besides, the second wait (on SETDISPLAY), which
was intended to avoid the possibility of presenting any contents that
were still half-rendered, now actually waits for the last render to
happen and we see no flickering. Before it was waiting for the wrong
event to happen leading to both: unneeded stalls and missed waits.

With the current patch and the needed HWC change for using the new
GEM_WAIT flag, the forementioned benchmark gets again its original
score and there is no flickering.

Change-Id: Ide67650754a1ca2d3d891b0ff81e29d525410fdb
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-4406
Signed-off-by: Abel Bernabeu <abel.bernabeu@intel.com>
Reviewed-on: https://android.intel.com:443/235369
Reviewed-on: https://android.intel.com:443/285496