x86/apic: Plug racy xAPIC access of CPU hotplug code
there could be migration/irq between writing APIC_ICR2 and APIC_ICR and
cause a wrong destination of INIT/SIPI flow.

This patch is a backport from upstream

apic_icr_write() and its users in smpboot.c were apparently
written under the assumption that this code would only run
during early boot. But nowadays we also execute it when onlining
a CPU later on while the system is fully running. That will make
wakeup_cpu_via_init_nmi and, thus, also native_apic_icr_write
run in plain process context. If we migrate the caller to a
different CPU at the wrong time or interrupt it and write to
ICR/ICR2 to send unrelated IPIs, we can end up sending INIT,
SIPI or NMIs to wrong CPUs.

Fix this by disabling interrupts during the write to the ICR
halves and disable preemption around waiting for ICR
availability and using it.

Original-signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>
Original-tested-by: Igor Mammedov <imammedo@redhat.com>
Link: http://lkml.kernel.org/r/52E6AFFE.3030004@siemens.co
Original-signed-off-by: Ingo Molnar <mingo@kernel.org>
Original-commit-id: ea7bdc65bca8cf837a63e0ff7b75daed83222511

Change-Id: I1949f904f189dde2633d8a5955702a20eb53c1ff
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-9983
Signed-off-by: Di Zhang <di.zhang@intel.com>
Signed-off-by: xiao jin <jin.xiao@intel.com>
Signed-off-by: wang, biao <biao.wang@intel.com>
Reviewed-on: https://android.intel.com:443/303742