drm/i915: Split set_display into val/calc and commit
As a preparation to "true" display atomic update, split the display
change analysis and actual write into hardware into different logical
sections
RRB2 is taken care as part of the flip operation.
Z-order is calculated and the cached in the register, written at the
begining of the commit and in the later places where sprite control
register is touched the first 3 bits related to z-order are non-masked.

Change-Id: Iabe1fd945e03f3f4d83535d38c21de7ba778dcd6
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-3947
Signed-off-by: Arun R Murthy <arun.r.murthy@intel.com>
Signed-off-by: Vijay Purushothaman <vijay.a.purushothaman@intel.com>
Reviewed-on: https://android.intel.com:443/285464