drm/i915: Enable/Disable Maxfifo
When moving from single plane to multiple plane scenario, before
enabling the second plane, disable maxfifo and wait for a vblank.
Similarly when in maxfifo display/sprite control register cannot be
changed, in that case, disable maxfifo wait for a vblank to get affect
and then write to the control register.

Change-Id: I328362fb444258f7ef4dce8e09e905f52bb4946d
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-3947
Signed-off-by: Arun R Murthy <arun.r.murthy@intel.com>
Signed-off-by: Vijay Purushothaman <vijay.a.purushothaman@intel.com>
Signed-off-by: Arun R Murthy <arun.r.murthy@intel.com>
Reviewed-on: https://android.intel.com:443/285488