drm/i915: DDL calculate and write to reg
DDL register is a single buffer register, hence any write to the
register will get into affect immediately. So during calculation if
there is a change in the calculated ddl value then write a value of '0'
to the DDL register and program the proper calculated value on the next
vblank, i.e in the flip done interrupt handler.

Change-Id: Ic5155b8fb9c225cd4ff80556ad33e0110c18e254
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-3947
Signed-off-by: Arun R Murthy <arun.r.murthy@intel.com>
Signed-off-by: Vijay Purushothaman <vijay.a.purushothaman@intel.com>
Signed-off-by: Arun R Murthy <arun.r.murthy@intel.com>
Reviewed-on: https://android.intel.com:443/285489