EM/Charger: Fix the klockwork issue
In memcpy, source string length is less than is smaller than destination
buffer but the size parameter is the maximum size of destination buffer.
Because of this, array indexes after the actual data could be used and could
potentially corrupt the memory.

This patch fixes the issue by copying only the actual size of the source string.

Change-Id: Ib5c121c3492fcbb7c1021fefb4a199c9a0a98bd5
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-4326
Signed-off-by: Rapaka Naveen <naveen.rapaka@intel.com>
Signed-off-by: Pandey Raj <raj.pandey@intel.com>
Reviewed-on: https://android.intel.com:443/285240