EM/THERMAL/MOOR: Disable SoC & CPU DTS interrupts
[Description]
To avoid prochot during heavy workloads,
a new mode called TP mode is being enabled
in Moorefield.
This requires the existing handling from
SW to be disabled

[Implementation]
To disable the existing SW handling, the
CPU and SoC DTS interrupts need to be disabled.
CPU DTS interrupts are disabled by disabling
the flag "SENSORS_CORETEMP_INTERRUPT"

Similalry, to disable the SoC DTS interrupts
a new config flag "SENSORS_SOCDTS_INTERRUPT"
is added and relevant interrupt code is moved
under this. This flag is enabled in the
defconfig files of all platforms and
disabled for moorefield.

Change-Id: I5dfd1f9d42e9a1b41c70df052a83f611a9e1699a
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-7998
Signed-off-by: Ajay Thomas <ajay.thomas.david.rajamanickam@intel.com>
Reviewed-on: https://android.intel.com:443/298320