drm/i915: Fix for flicker caused by drmModeSetDisplay stalls
The page flip call that generates the flip complete callback was
being issued as the first page flip in the kernel. Hence, if anything
caused the setPlane calls to stall, then the HWC could end up indicating
the page flip event had completed prior to the whole frame having been
updated.
This commit reorders the flips such that the pageflip is last and hence
the flip complete event will happen after all the setPlane calls.

Change-Id: I1b80c6e24e51bc129cf13e5e9e045df9004b2b1f
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-8669
Signed-off-by: Gary Smith <gary.k.smith@intel.com>
Signed-off-by: Animesh Manna <animesh.manna@intel.com>
Reviewed-on: https://android.intel.com/#/c/247834/
Reviewed-on: https://android.intel.com:443/299446