update config_name to support 7160 in user build
On CTP platform, we are relying on an engineering mechanism to
support the scalability and have the right CONF name.
But this mechanism is not possible on user build for security
reasons. The CONF name MUST be written in the sysfs and it is
provided by the SFI table.
Because we cannot update the SFI content without forcing a
blankphone, we need a hack in the kernel to catch the SFI content and
write the new CONF name instead.

Change-Id: Id59de3e979288d69bc09a22ffd83962c8dcdecb9
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-10638
Signed-off-by: Ovidiu Constantin Mitocaru <ovidiu.constantinx.mitocaru@intel.com>
Reviewed-on: https://android.intel.com:443/304555