[RAMDUMP] Correct an issue in x86 mode
Correct a crash due to wrong parameter in case of x86 arch:

If we look at Interrupt handler asm code (Entry_32.S), we can see that it
passes pt_regs thru eax :

movl %esp,%eax;
call fn;

If we look at kernel/arch/x86/Makefile, we can see:
KBUILD_CFLAGS += -msoft-float -mregparm=3 -freg-struct-return

"-mregparm=3" means that 3 registers will be used to pass arguments.

The "asmlinkage" tag overwrites this and forces arguments to be found in
stack. This is not compliant with Handler code.

Change-Id: I0521286f4b66473266e8a74b1140d41abf5d8063
Signed-off-by: Emmanuel Berthier <emmanuel.berthier@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-12436
Orig-Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-11139
Reviewed-on: https://android.intel.com:443/309005